+++ tavern/_core/starlark/__init__.py
@@ -1,7 +1,6 @@
 """Starlark pipeline support for Tavern."""
 
 from .pytest_plugin import pytest_collect_file
-from .runner import StageResponse, run_stage
 from .starlark_env import (
     PipelineContext,
     StarlarkPipelineRunner,
@@ -9,8 +8,6 @@ from .starlark_env import (
 
 __all__ = [
     "PipelineContext",
-    "StageResponse",
     "StarlarkPipelineRunner",
     "pytest_collect_file",
-    "run_stage",
 ]
